requirements.txt
pandas
setuptools
click
pytest
numpy
environs
ipython
markovify
python-telegram-bot
telethon
tqdm
mega.py
